Protect task running with tryCatch
This will not always do the right thing; if there is an internal error (such as the task going missing) then we'll get the wrong error message (not the error
from running the task but the error from the general system).  But in general
this should stop the worker dying.