Standardise pluralisation
This totally breaks the package (again) but basically nobody is
depending on it yet apart from Pete so I can work around that for
now.

This rolls back the "plural verb" approach that I used here and in
queuer, but proved confusing for non-native English speakers.  Most
things can be done with a singular verb everywhere with relatively
few special cases.  There is some task result bits left to deal
with but these are difficult because of type safety.