Propagate the offload path to sub-processes.
When a task is run in a separate process, a new rrq_worker object and
controller are created in the sub-process. These need to be configured
to use the right offload directory.

It's not obvious to me that the sub-process even needs to interact with
Redis and the offload directory. It could very well let the main worker
process load the task data and receive it in its arguments, letting
callr handle the transfer. In fact that's already how the result is
handled, it get returned by the subprocess and written to the offload
storage by the parent process.